jdbc/src/main/java/org/springframework/jdbc/core/PreparedStatementQueryExecuter.java
package org.springframework.jdbc.core;

import java.sql.PreparedStatement;
import java.sql.ResultSet;
import java.sql.SQLException;
import java.util.ArrayList;
import java.util.List;

public class PreparedStatementQueryExecuter<T> implements PreparedStatementExecuter<List<T>> {

    private final RowMapper<T> rowMapper;

    public PreparedStatementQueryExecuter(RowMapper<T> rowMapper) {
        this.rowMapper = rowMapper;
    }

    @Override
    public List<T> execute(PreparedStatement pstmt) throws SQLException {
        try (ResultSet resultSet = pstmt.executeQuery()) {
            List<T> results = new ArrayList<>();
            while (resultSet.next()) {
                results.add(rowMapper.mapRow(resultSet));
            }
            return results;
        }
    }
}
